+++ R/cli.R
@@ -46,6 +46,7 @@ retrieve_source <- function(source_id, refresh = FALSE, max_age = NULL) {
     orwn_station = retrieve_orwn_station(refresh = refresh),
     monitoring_stations = retrieve_monitoring_stations(refresh = refresh),
     monitoring_stations_simple = retrieve_monitoring_stations_simple(),
+    postal_points = retrieve_postal_points(refresh = refresh),
     {
       valid_ids <- list_sources()$source_id
       if (source_id %in% valid_ids) {
@@ -285,6 +286,7 @@ source_retrieve_call <- function(source_id) {
     orwn_station = "retrieve_orwn_station()",
     monitoring_stations = "retrieve_monitoring_stations()",
     monitoring_stations_simple = "retrieve_monitoring_stations_simple()",
+    postal_points = "retrieve_postal_points()",
     synthetic_air_quality = "retrieve_synthetic_raster()",
     hive = "retrieve_hive()",
     {